+++ src/typeguard/_checkers.py
@@ -49,6 +49,7 @@ from ._utils import evaluate_forwardref, get_stacklevel, get_type_name, qualifie
 if sys.version_info >= (3, 11):
     from typing import (
         NotRequired,
+        Required,
         TypeAlias,
         get_args,
         get_origin,
@@ -59,6 +60,7 @@ else:
     from typing_extensions import Any as SubclassableAny
     from typing_extensions import (
         NotRequired,
+        Required,
         TypeAlias,
         get_args,
         get_origin,
@@ -266,6 +268,9 @@ def check_typed_dict(
 
         if get_origin(annotation) is NotRequired:
             required_keys.discard(key)
+            annotation = get_args(annotation)[0]
+        elif get_origin(annotation) is Required:
+            required_keys.add(key)
             annotation = get_args(annotation)[0]
 
         type_hints[key] = annotation